refactor: split collect_data pipeline into collectors and build modules
Move each scraper into ``collect_data/collectors`` (fbref, understat,
sofascore, sofascore_matches, clubelo, transfermarkt, capology) and merge
logic into ``collect_data/build`` (``unified.py``, ``financials.py``). Keep
``pipeline.py`` as CLI-only orchestration. Add ``fbref_season_to_understat`` to
helpers for Understat. Update the README structure diagram and public exports
so tests and ``python -m collect_data`` behave as before.
