+++ collect_data/__init__.py
@@ -9,12 +9,10 @@ or the compatibility script ``collect_data.py`` in the repository root.
 
 from __future__ import annotations
 
+from collect_data.build.unified import build_unified
+from collect_data.collectors.sofascore_matches import _sofascore_match_pack_fully_done
 from collect_data.config import SEASONS
-from collect_data.pipeline import (
-    _sofascore_match_pack_fully_done,
-    build_unified,
-    main,
-)
+from collect_data.pipeline import main
 
 __all__ = [
     "SEASONS",
